Move deep_align back into alignment.py
