+++ xarray/core/alignment.py
@@ -25,8 +25,11 @@ def _get_joiner(join):
         raise ValueError('invalid value for join: %s' % join)
 
 
+_DEFAULT_EXCLUDE = frozenset()
+
 def align(*objects, **kwargs):
-    """align(*objects, join='inner', copy=True)
+    """align(*objects, join='inner', copy=True, indexes=None,
+             exclude=frozenset())
 
     Given any number of Dataset and/or DataArray objects, returns new
     objects with aligned indexes.
@@ -66,11 +69,9 @@ def align(*objects, **kwargs):
     join = kwargs.pop('join', 'inner')
     copy = kwargs.pop('copy', True)
     indexes = kwargs.pop('indexes', None)
-    exclude = kwargs.pop('exclude', None)
+    exclude = kwargs.pop('exclude', _DEFAULT_EXCLUDE)
     if indexes is None:
         indexes = {}
-    if exclude is None:
-        exclude = set()
     if kwargs:
         raise TypeError('align() got unexpected keyword arguments: %s'
                         % list(kwargs))
@@ -117,12 +118,68 @@ def align(*objects, **kwargs):
     return tuple(result)
 
 
+def deep_align(objects, join='inner', copy=True, indexes=None,
+               exclude=frozenset(), raise_on_invalid=True):
+    """Align objects for merging, recursing into dictionary values.
+
+    This function is not public API.
+    """
+    if indexes is None:
+        indexes = {}
+
+    def is_alignable(obj):
+        return hasattr(obj, 'indexes') and hasattr(obj, 'reindex')
+
+    positions = []
+    keys = []
+    out = []
+    targets = []
+    no_key = object()
+    not_replaced = object()
+    for n, variables in enumerate(objects):
+        if is_alignable(variables):
+            positions.append(n)
+            keys.append(no_key)
+            targets.append(variables)
+            out.append(not_replaced)
+        elif is_dict_like(variables):
+            for k, v in variables.items():
+                if is_alignable(v) and k not in indexes:
+                    # Skip variables in indexes for alignment, because these
+                    # should to be overwritten instead:
+                    # https://github.com/pydata/xarray/issues/725
+                    positions.append(n)
+                    keys.append(k)
+                    targets.append(v)
+            out.append(OrderedDict(variables))
+        elif raise_on_invalid:
+            raise ValueError('object to align is neither an xarray.Dataset, '
+                             'an xarray.DataArray nor a dictionary: %r'
+                             % variables)
+        else:
+            out.append(variables)
+
+    aligned = align(*targets, join=join, copy=copy, indexes=indexes,
+                    exclude=exclude)
+
+    for position, key, aligned_obj in zip(positions, keys, aligned):
+        if key is no_key:
+            out[position] = aligned_obj
+        else:
+            out[position][key] = aligned_obj
+
+    # something went wrong: we should have replaced all sentinel values
+    assert all(arg is not not_replaced for arg in out)
+
+    return out
+
+
 def reindex_variables(variables, indexes, indexers, method=None,
                       tolerance=None, copy=True):
     """Conform a dictionary of aligned variables onto a new set of variables,
     filling in missing values with NaN.
 
-    WARNING: This method is not public API. Don't use it directly.
+    This function is not public API.
 
     Parameters
     ----------
